Version 2.1
Extract the class “PickleData”, out of the Controller class and place in a new file
