+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assessment 2 Initial\\unit_tests.py":[2,3,4,5,6,7,8,9,10,11,14,15,22,32,97,106,115,122,128,136,147,158,166,173,181,205,223,238,248,258,268,277,286,299,306,313,323,338,347,356,365,374,384,391,398,405,412,418,423,431,439,447,455,463,471,478,486,495,496,16,17,18,19,20,182,183,184,185,24,25,26,27,28,186,30,187,188,203,224,225,226,227,228,229,230,236,239,240,241,242,243,244,245,246,249,250,251,252,253,254,255,256,259,260,261,262,263,264,265,266,269,270,271,272,273,274,275,278,279,280,281,282,283,284,287,288,289,290,297,472,473,474,475,476,487,488,489,490,491,492,33,34,35,36,37,95,385,386,387,388,389,392,393,394,395,396,399,400,401,402,403,406,407,408,409,410,314,315,316,317,318,319,320,321,324,325,326,327,336,339,340,341,342,343,344,345,348,349,350,351,352,353,354,357,358,359,360,361,362,363,366,367,368,369,370,371,372,375,376,377,378,379,380,381,382,206,207,208,209,210,211,212,221,174,175,176,177,178,179,148,149,150,151,152,153,156,159,160,161,162,163,164,167,168,169,170,171,137,138,139,140,141,142,145,419,420,421,413,414,415,416,98,99,100,101,102,104,123,124,125,126,107,108,109,110,111,113,116,117,118,119,120,129,130,131,132,133,134,456,457,458,459,461,464,465,466,467,469,440,441,442,443,445,432,433,434,435,437,424,425,426,427,429,448,449,450,451,453,479,480,481,482,484,300,301,302,303,304,307,308,309,310,311],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assessment 2 Initial\\cmd_view.py":[2,6,8,13,17,20,25,28,34,37,42,45,50,53,58,63,68,71,9,10,11,14,35,43,18,51,60,61,69,26,55,65,39,30,31,22,47,73],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assessment 2 Initial\\controller.py":[2,3,4,5,6,7,8,9,12,13,25,31,36,42,64,90,116,144,154,163,14,15,16,17,18,19,20,21,22,23,91,92,93,94,95,96,97,98,145,146,147,148,65,66,68,69,70,71,72,73,117,118,119,120,121,122,123,124,125,128,134,142,129,130,131,127,133,135,136,137,138,139,140,156,157,158,99,100,108,110,111,113,114,112,43,44,45,46,47,33,34,48,51,52,53,38,39,40,54,55,56,50,58,59,61,62,60,74,75,77,78,79,80,81,82,83,84,85,86,67,88,76,26,27,28,29,150,151,152,101,102,103,104,105,106,160],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assessment 2 Initial\\graph_maker.py":[2,3,6,7,8,9,10,11,12,13,15,17,20],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assessment 2 Initial\\database_view.py":[2,3,4,7,8,14,19,37,43,9,10,11,12,20,15,16,17,21,22,23,30,31,32,33,34,35,38,39,40,41,44,45,46,47,48,50,51,53,49,54,55],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assessment 2 Initial\\validate.py":[2,3,6,7,17,32,37,46,53,57,8,10,11,12,13,14,15,58,59,60,61,62,63,64,65,71,72,73,91,48,49,92,97,102,74,34,35,75,80,81,84,86,19,20,21,22,25,30,87,39,40,41,42,103,104,51,98,99,101,67,68,69,76,77,106,108,82,83,26,28,29,89,93,94,96,78,44,27,23,55,85],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assessment 2 Initial\\file_entry_view.py":[2,5,6,9,14,17,7,10,11,12,15]}}
+!coverage.py: This is a private format, don't read it directly!{"lines":{"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assignment Two Version Control\\Version 2 - Long Class\\Assessment 2 - Version 1.1\\unit_tests.py":[2,3,4,5,6,7,8,9,10,11,14,15,22,32,97,106,115,122,128,136,147,158,166,173,181,205,223,238,248,258,268,277,286,299,306,313,323,338,347,356,365,374,384,391,398,405,412,418,423,431,439,447,455,463,471,478,486,495,496,16,17,18,19,20,182,183,184,185,24,25,26,27,28,186,30,187,188,203,224,225,226,227,228,229,230,236,239,240,241,242,243,244,245,246,249,250,251,252,253,254,255,256,259,260,261,262,263,264,265,266,269,270,271,272,273,274,275,278,279,280,281,282,283,284,287,288,289,290,297,472,473,474,475,476,487,488,489,490,491,492,33,34,35,36,37,95,385,386,387,388,389,392,393,394,395,396,399,400,401,402,403,406,407,408,409,410,314,315,316,317,318,319,320,321,324,325,326,327,336,339,340,341,342,343,344,345,348,349,350,351,352,353,354,357,358,359,360,361,362,363,366,367,368,369,370,371,372,375,376,377,378,379,380,381,382,206,207,208,209,210,211,212,221,174,175,176,177,178,179,148,149,150,151,152,153,156,159,160,161,162,163,164,167,168,169,170,171,137,138,139,140,141,142,145,419,420,421,413,414,415,416,98,99,100,101,102,104,123,124,125,126,107,108,109,110,111,113,116,117,118,119,120,129,130,131,132,133,134,456,457,458,459,461,464,465,466,467,469,440,441,442,443,445,432,433,434,435,437,424,425,426,427,429,448,449,450,451,453,479,480,481,482,484,300,301,302,303,304,307,308,309,310,311],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assignment Two Version Control\\Version 2 - Long Class\\Assessment 2 - Version 1.1\\cmd_view.py":[2,6,8,13,17,20,25,28,34,37,42,45,50,53,58,63,68,71,9,10,11,14,35,43,18,51,60,61,69,26,55,65,39,30,31,22,47,73],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assignment Two Version Control\\Version 2 - Long Class\\Assessment 2 - Version 1.1\\controller.py":[2,3,4,5,6,7,8,9,12,13,26,32,37,63,89,117,127,136,14,15,16,17,18,19,20,21,22,23,24,64,65,66,67,68,69,70,71,118,119,120,121,38,39,41,42,43,44,45,46,90,91,92,93,94,95,96,97,98,101,107,115,102,103,104,100,106,108,109,110,111,112,113,129,130,131,72,73,81,83,84,86,87,85,33,34,35,47,48,50,51,52,53,54,55,56,57,58,59,40,61,49,27,28,29,30,123,124,125,74,75,76,77,78,79,133],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assignment Two Version Control\\Version 2 - Long Class\\Assessment 2 - Version 1.1\\graph_maker.py":[2,3,6,7,8,9,10,11,12,13,15,17,20],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assignment Two Version Control\\Version 2 - Long Class\\Assessment 2 - Version 1.1\\pickle_data.py":[2,5,7,11,39,44,50,8,9,12,13,14,15,16,17,18,41,42,19,51,23,24,25,46,47,48,26,27,28,21,30,32,35,36,33],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assignment Two Version Control\\Version 2 - Long Class\\Assessment 2 - Version 1.1\\database_view.py":[2,3,4,7,8,14,19,37,43,9,10,11,12,20,15,16,17,21,22,23,30,31,32,33,34,35,38,39,40,41,44,45,46,47,48,50,51,53,49,54,55],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assignment Two Version Control\\Version 2 - Long Class\\Assessment 2 - Version 1.1\\validate.py":[2,3,6,7,17,32,37,46,53,57,8,10,11,12,13,14,15,58,59,60,61,62,63,64,65,71,72,73,91,48,49,92,97,102,74,34,35,75,80,81,84,86,19,20,21,22,25,30,87,39,40,41,42,103,104,51,98,99,101,67,68,69,76,77,106,108,82,83,26,28,29,89,93,94,96,78,44,27,23,55,85],"D:\\Documents\\CPIT\\2017\\Advanced Programming\\Bad Smells Refactoring Code\\Assignment Two Version Control\\Version 2 - Long Class\\Assessment 2 - Version 1.1\\file_entry_view.py":[2,5,6,9,14,17,7,10,11,12,15]}}
